- moved all german const to language german - moved other constants to constants.rs - fmt
