+++ game/src/backend/constants.rs
@@ -1,0 +1,8 @@
+/// This file contains constants that are necessary for the game.
+
+/// Contains the map border( x-right, y-bottom, x-left, y-top)
+pub const MAP_BORDER: [usize; 4] = [1750, 850, 255, 220];
+
+// pub const MACHINE_POSITIONS: [[i32; 4]; 4] = [[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0]];
+
+// maybe add rl color here
