+++ game/src/backend/mod.rs
@@ -1,10 +1,10 @@
 pub(crate) mod area;
 pub(crate) mod asset_service;
+mod constants;
 pub(crate) mod error;
 pub(crate) mod gamestate;
 pub(crate) mod generate_machines;
 pub(crate) mod movement;
-pub(crate) mod popup_messages;
 pub(crate) mod rlcolor;
 pub(crate) mod screen;
 pub(crate) mod utils;
