--- game/src/backend/popup_messages.rs
@@ -1,22 +1,0 @@
-/// This is the text for the popups that appears in the top left corner.
-
-pub const WARNINGS: [&str; 4] = [
-    "Ein Komet ist auf dem Weg!",
-    "Der Strom ist ausgefallen!",
-    "Ein Sandsturm ist auf dem Weg!",
-    "Eine Maschine ist ausgefallen!",
-];
-pub const MARS_INFO: [&str; 5] = [
-    "Der Mars ist der 4. Planet in unserem Sonnensystem",
-    "Der Mars zählt zu den erdähnlichen Planeten",
-    "Der Durchmesser des Mars beträgt knapp 6800 km",
-    "Die Masse des Mars beträgt etwa ein Zehntel der Erdmasse",
-    "Die Entfernung zum Mars beträgt durchschnittlich 228 Millionen km",
-];
-pub const NASA_INFO: [&str; 5] = [
-    "NASA steht für: National Aeronautics and Space Administration",
-    "Die NASA wurde 1958 gegründet",
-    "Die NASA hat ihren Sitz in Washington D.C.",
-    "Im Rahmen der Apollo-Missionen gelang es der NASA, den ersten Menschen auf den Mond zu bringen"," Die NASA hat über 17.000 Mitarbeiter"
-];
-pub const GAME_INFO: [&str; 1] = ["Lebensregeneration gestartet"];
