+++ game/src/languages/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod german;
